Use Long IDs, rename params; add front models
Change Plan-related ID parameters from int to Long (PlanRepository, PlanService, PlanServiceImpl) to match entity types and avoid type mismatches. Update multiple controllers (Company, Loyalty, Plan, Reward, Store, User) to use explicit @PathVariable names and renamed local variables for clarity; remove PreAuthorize on StoreController.findAll and adjust several method signatures accordingly. Add frontend Angular model interfaces and services (Company, LoyaltyCard, Plan, Reward, Store, Transaction, User, environment) plus AuthService and basic unit spec files to support API integration.

+++ backend/src/main/java/tienda/puntos/app/repository/dao/PlanRepository.java
@@ -12,9 +12,9 @@ import tienda.puntos.app.repository.entity.Plan;
 public interface PlanRepository extends JpaRepository<Plan, Long> {
 
     @Query(value = "UPDATE plans SET active = 0 WHERE id = ?1", nativeQuery = true)
-    void disable(int planId);
+    void disable(Long planId);
 
     @Query(value = "UPDATE plans SET active = 1 WHERE id = ?1", nativeQuery = true)
-    void active(int planId);
+    void active(Long planId);
 
 }
